requirements.txt
certifi==2022.9.24
click==8.1.3
dnspython==2.2.1
Flask==2.2.2
Flask-Cors==3.0.10
gunicorn==20.1.0
importlib-metadata==5.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
pymongo==4.2.0
python-dotenv==0.21.0
six==1.16.0
Werkzeug==2.2.2
zipp==3.8.1
